wl12xx: use MAC address from the fuse if not available in the NVS
Newer PG versions contain the BD_ADDR in the fuse ROM.  Beside the
BD_ADDR, the next two addresses are reserved to WLAN.  If the NVS
doesn't have a valid MAC address (ie. if it is set to 00:00:00:00:00),
we read the BD_ADDR from the fuse ROM and derive the WLAN MAC
addresses from it.

Signed-off-by: Luciano Coelho <coelho@ti.com>