wl12xx: check for vif state instead of device state
During recovery there is a race in which the device might
be initialized but the vif is not, which finally enqueues
a scan request on a vif which was already deinitialized,
and prevents further scans from being initiated.

fix it by checking the vif state instead of the device state.

TODO: this should probably be done in some more places

Signed-off-by: Eliad Peller <eliad@wizery.com>