wl12xx: WiFi doesn't auto connect
We previously didn't wait for this event, as it wasn't
necessary, and sometimes the FW didn't send it.
However, this event is now needed again for some roaming
scenarios.

If the driver received a watchdog interrupt then the
assumption is that the FW is hanged. Avoid sending
the stop FW log command in case of a watchdog recovey
to avoid waiting for the 2 seconds timeout of the command.