wl12xx: Wait a little before retrying null frame
When entering PS we try to send the null frame 8 times. If that fails
we'll notify mac80211 of disconnection. The retry mechanism would
send a new nullfunc directly after the previous failed so all 8 null
frames are fired of in about 40ms. If we have some intermittent
interference or the AP is busy we can easily get disconnected.

Fix this be adding a delay of 20 ms before resending the null frame.
Also change number of retires to 20. This means we'll retry for
a period of 400 ms. 400 was chosen not to interfer with the 500 ms
timeout in system suspend.

Category: bugfix
FIX=DMS01133469
Change-Id: Ibf3aa2de93cf85c0ba84fc38c9875b46ea824fd6

Signed-off-by: Pontus Fuchs <pontus2.fuchs@sonymobile.com>