mac80211: Remove the floor scan threshold
for -88dBm. With this floor defined and used, AP scan results
are less than that from GB, since that platform did not use
such cutoff.

WiFi-BT coex roaming issues were previously addressed with
WLAN firmware and driver fixes, so this floor threshold is not
required to improve prior poor performance.