mac80211: use legacy mode for all ACs
FritzBox 7390 (and potentially other models) have a weird behaviour
where once we enable U-APSD on the the VO AC it limits the rate
of the packets it sends us to 39Mbps and doesn't do AMPDU aggregation.
This causes a big degradation in TP obviously.
Once we use legacy mode for all ACs it's willing to use higher rates
as well as do AMPDU aggregation.

Signed-off-by: Eyal Shapira <eyal@wizery.com>