wl12xx: Check return value when reading the FW status
The return value of the bus read operation is not checked when reading
the FW status. This can lead to various problems. For instance, the
driver might not be able to acknowledge interrupts, which will keep on
firing as soon as the interrupt handler exits. Fix this by propagating
the bus operation return value and handling it.

Signed-off-by: Ido Yariv <ido@wizery.com>