wl12xx: init the pll reg before reading the fuse reg
The device mac address can be stored in the nvs or
the fuse registers. If the nvs contains a zero mac
address then the driver will fallback to the fuse
mac address which is read during the wl12xx_sdio
insmod.

The fuse registers are held in the DRPW partition,
therefore in order to read them safely the pll
clock must first be initialized.

Signed-off-by: Yoni Divinsky <yoni.divinsky@ti.com>