wl12xx: always clear recovery flag during recovery_work
If recovery is called when the FW is off, we should clear
the recovery flag. Otherwise we risk booting the driver in
permanent pending-recovery state or getting into a state where
the driver is left in an OFF state.
This can happen if a 2nd recovery is triggered and runs before
ieee80211_reconfig adds back the interface.
Fix this by always clearing the recovery flag.
In addition, enable interrupts to balance the disable in
wl1271_queue_recovery_work.

(This is based on Arik's patch for wlcore with fixes)

Signed-off-by: Eyal Shapira <eyal@wizery.com>